% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata.download.R
\name{bivariate_download}
\alias{bivariate_download}
\title{The bivariate_download function}
\usage{
bivariate_download(
  base,
  dir,
  sub_path,
  number_factors,
  subtype,
  freq,
  dividends,
  start,
  end,
  clean_na
)
}
\arguments{
\item{base}{a character string,
the main webpage address "http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/".}

\item{dir}{a character string, the directory for saving the data. The current working directory is set by default.}

\item{sub_path}{a character string, the folder subpath, created by ffDataDownload.}

\item{number_factors}{an integer, the number of factors for the Bivariate portfolios.
Possible values are 6 (default), 25 or 100.}

\item{subtype}{a character string, the Bivariate portfolio subtype.
Possible values are "ME_BE" (default), "ME_OP", "ME_INV", "BEME_OP", "BEME_INV", "OP_INV".}

\item{freq}{a character string, the frequency of the returns, "m" for monthly (default) or "d" for daily.}

\item{dividends}{a logical, TRUE (default) considers dividends. FALSE downloads the data without dividends.}

\item{start}{a character string, start date for the download in the format "YYmm". Default value is 197501.}

\item{end}{a character string, end date for the download in the format "YYmm".
Default value is two months before Sys.Date() to insure availability.}

\item{clean_na}{a logical, TRUE (default) replaces NAs with zoo::na.locf(). If FALSE, NAs are not cleaned.}
}
\value{
a .csv-file within the directory, defined in dir.
}
\description{
This function automatically downloads Bivariate portfolios returns
from the Kenneth R. French Data Library.
The downloaded returns are saved then as a .csv-file in the chosen directory.
}
