% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata.download.R
\name{ffdata.download}
\alias{ffdata.download}
\title{The ffdata.download function}
\usage{
ffdata.download(dir = NULL, type = "USResearch", subtype = NULL,
  factors.n = NULL, freq = "m", dividends = TRUE, start = NULL,
  end = NULL, na.clean = TRUE)
}
\arguments{
\item{dir}{character, the directory for saving the data. If not specified, current working directory (default).}

\item{type}{character, the portfolio type, "USResearch" (default) for the Fama-French 3 or 5 Factors,
"Industry" for the Industry portfolios, "Bivariate" for the Bivariate sorts on Size (ME), Book-to-Market (BE), Operational Profitability (OP) and Investment (INV),
"Threeway" for the Three-way sorts on Size (ME), Book-to-Market (BE), Operational Profitability (OP) and Investment (INV).}

\item{subtype}{character, the portfolio subtype. Possible values for the Bivariate type: "ME_BE", "ME_OP", "ME_INV", "BEME_OP", "BEME_INV", "OP_INV", possible
values for the Threeway type: "ME_BEME_OP", "ME_BEME_INV", "ME_OP_INV".}

\item{factors.n}{integer, the number of factors for the portfolios. Possible values for the USResearch type: 3, 5. Possible values for the Industry type: 5, 10, 12, 17, 30, 38, 48, 49.
Possible values for the Bivariate type: 6, 25, 100.}

\item{freq}{character, the frequency of the returns, "m" for monthly (default) or "d" for daily.}

\item{dividends}{logical, with dividends (default) if TRUE and without dividends if FALSE.}

\item{start}{character, start date for the download in the format YYmm (default).}

\item{end}{character, end date for the download in the format YYmm. The default value is calculated as two months before Sys.Date() to reassure availability.}

\item{na.clean}{logical, if TRUE, NAs are replaced with zoo::na.locf() (default). If FALSE, NAs are not cleaned.}
}
\value{
.csv-file within the defined directory.
}
\description{
This function performs the automatic download of the user-specified portfolio returns from the Kenneth R. French Data Library.
The downloaded data frame of returns is saved then as a .csv-file in the chosen directory.
}
\examples{
ffdata.download()
ffdata.download(freq="m", type="Bivariate", subtype="ME_BE", factors.n=25)
}
