% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata.download.R
\name{ffdata_download}
\alias{ffdata_download}
\title{The ffdata_download function}
\usage{
ffdata_download(
  dir = NULL,
  type = "USResearch",
  subtype = NULL,
  number_factors = NULL,
  freq = "m",
  dividends = TRUE,
  start = NULL,
  end = NULL,
  clean_na = TRUE
)
}
\arguments{
\item{dir}{a character string, the directory for saving the data.
The current working directory is set by default.}

\item{type}{a character string, the portfolio type.
Possible values are "USResearch" (default) for the Fama-French 3 or 5 Factors,
"Industry" for the Industry portfolios, "Bivariate" for the Bivariate sorts on Size (ME),
Book-to-Market (BE), Operational Profitability (OP) and Investment (INV),
"Threeway" for the ME, BE, OP, and INV.}

\item{subtype}{a character string, the portfolio subtype.
Possible values are "ME_BE", "ME_OP", "ME_INV", "BEME_OP", "BEME_INV", "OP_INV" (for the Bivariate type) and
"ME_BEME_OP", "ME_BEME_INV", "ME_OP_INV" (for the Threeway type).}

\item{number_factors}{an integer, the number of factors for the portfolios.
Possible values are 3 (default) or 5 (for the USResearch portfolios);
5 (default), 10, 12, 17, 30, 38, 48 or 49 (for the Industry portfolios) and
6 (default), 25 or 100 (for the Bivariate portfolios).}

\item{freq}{a character string, the frequency of returns.
Possible values are "m" for monthly (default) or "d" for daily.}

\item{dividends}{a logical, TRUE (default) considers dividends. FALSE downloads the data without dividends.}

\item{start}{a character string, start date for the download in the format "YYmm". Default value is 197501.}

\item{end}{a character string, end date for the download in the format "YYmm".
Default value is two months before Sys.Date() to insure availability.}

\item{clean_na}{a logical, TRUE (default) replaces NAs with zoo::na.locf(). If FALSE, NAs are not cleaned. # nolint: line_length_linter.}
}
\value{
a .csv-file within the directory, defined in dir.
}
\description{
This function automatically downloads user-specified portfolio returns from
the Kenneth R. French Data Library.
The downloaded returns are saved as a .csv-file in the chosen directory.
}
\examples{
library(ffdata)
ffdata_download()
ffdata_download(freq = "m", type = "Bivariate", subtype = "ME_BE", number_factors = 25)
}
