% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata.download.R
\name{industry.download}
\alias{industry.download}
\title{The industry.download function}
\usage{
industry.download(base, dir, sub.path, factors.n, freq, dividends, start,
  end, na.clean)
}
\arguments{
\item{base}{character string with the main webpage address "http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/".}

\item{dir}{character, the directory for saving the data. If not specified, current working directory (default).}

\item{sub.path}{character string with the folder subpath, created by ffdata.download.}

\item{factors.n}{integer, the number of factors for the portfolios. Possible values for the Industry type: 5, 10, 12, 17, 30, 38, 48, 49.}

\item{freq}{character, the frequency of the returns, "m" for monthly (default) or "d" for daily.}

\item{dividends}{logical, with dividends (default) if TRUE and without dividends if FALSE.}

\item{start}{character, start date for the download in the format YYmm (default).}

\item{end}{character, end date for the download in the format YYmm. The default value is calculated as two months before Sys.Date() to reassure availability.}

\item{na.clean}{logical, if TRUE, NAs are replaced with zoo::na.locf() (default). If FALSE, NAs are not cleaned.}
}
\value{
.csv-file within the defined directory.
}
\description{
This function performs the automatic download of the Industry portfolio returns from the Kenneth R. French Data Library.
The downloaded data frame of returns is saved then as a .csv-file in the chosen directory.
}
