% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata.download.R
\name{threeway.download}
\alias{threeway.download}
\title{The threeway.download function}
\usage{
threeway.download(base, dir, sub.path, subtype, dividends, start, end,
  na.clean)
}
\arguments{
\item{base}{character string with the main webpage address "http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/".}

\item{dir}{character, the directory for saving the data. If not specified, current working directory (default).}

\item{sub.path}{character string with the folder subpath, created by ffdata.download.}

\item{subtype}{character, the portfolio subtype. Possible values for the Threeway type: "ME_BEME_OP", "ME_BEME_INV", "ME_OP_INV".}

\item{dividends}{logical, with dividends (default) if TRUE and without dividends if FALSE.}

\item{start}{character, start date for the download in the format YYmm (default).}

\item{end}{character, end date for the download in the format YYmm. The default value is calculated as two months before Sys.Date() to reassure availability.}

\item{na.clean}{logical, if TRUE, NAs are replaced with zoo::na.locf() (default). If FALSE, NAs are not cleaned.}
}
\value{
.csv-file within the defined directory.
}
\description{
This function performs the automatic download of the Threeway portfolio returns from the Kenneth R. French Data Library.
The downloaded data frame of returns is saved then as a .csv-file in the chosen directory.
}
