% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdata_download.R
\name{threeway_download}
\alias{threeway_download}
\title{The threeway_download function}
\usage{
threeway_download(
  base,
  dir,
  sub_path,
  subtype,
  dividends,
  start,
  end,
  clean_na
)
}
\arguments{
\item{base}{a character string, the Kenneth R. French Data Library address.}

\item{dir}{a character string, the directory for saving the data. The current working directory is set by default.}

\item{sub_path}{a character string, the folder subpath, created by ffDataDownload.}

\item{subtype}{a character string, the Threeway portfolio subtype.
Possible values are "ME_BEME_OP" (default), "ME_BEME_INV" or "ME_OP_INV".}

\item{dividends}{a logical, TRUE (default) considers dividends. FALSE downloads the data without dividends.}

\item{start}{a character string, start date for the download in the format "YYmm". Default value is 197501.}

\item{end}{a character string, end date for the download in the format "YYmm".
Default value is two months before Sys.Date() to insure availability.}

\item{clean_na}{a logical, TRUE (default) replaces NAs with zoo::na.locf(). If FALSE, NAs are not cleaned.}
}
\value{
a .csv-file within the directory, defined in dir.
}
\description{
This function automatically downloads Threeway portfolios returns
from the Kenneth R. French Data Library.
The downloaded returns are saved then as a .csv-file in the chosen directory.
}
